% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{Heatmap}
\alias{Heatmap}
\title{rhtmlHeatmap widget}
\source{
The interface was designed based on \link{heatmap} and \link[gplots]{heatmap.2}
}
\usage{
Heatmap(x, Rowv = TRUE, Colv = if (symm) "Rowv" else TRUE, distfun = dist,
  hclustfun = hclust, dendrogram = c("both", "row", "column", "none"),
  reorderfun = function(d, w) reorder(d, w), k_row, k_col, symm = FALSE,
  revC, scale = c("none", "row", "column"), na.rm = TRUE,
  labRow = rownames(x), labCol = colnames(x), cexRow, cexCol, digits = 3L,
  cellnote, cellnote_scale = FALSE, cellnote_in_cell = NULL,
  show_cellnote_in_cell = FALSE, extra_tooltip_info = NULL,
  color_range = NULL, lower_triangle = FALSE, cells_to_hide = NULL,
  cell_font_size = 11, cell_font_family = "sans-serif",
  tip_font_size = 11, tip_font_family = "sans-serif", show_legend = TRUE,
  legend_font_size = 11, legend_font_family = "sans-serif",
  legend_font_color = "#000000", legend_width = 60, legend_digits = 1,
  theme = NULL, colors = "RdYlBu", width = NULL, height = NULL,
  title = NULL, title_font_size = 24, title_font_family = "sans-serif",
  title_font_color = "#000000", subtitle = NULL, subtitle_font_size = 18,
  subtitle_font_family = "sans-serif", subtitle_font_color = "#000000",
  footer = NULL, footer_font_size = 11, footer_font_family = "sans-serif",
  footer_font_color = "#000000", xaxis_hidden = FALSE, xaxis_height = 80,
  xaxis_font_size = 12, xaxis_font_family = "sans-serif",
  xaxis_font_color = "#000000", xaxis_location = "bottom",
  xaxis_title = NULL, xaxis_title_font_size = 14,
  xaxis_title_font_family = "sans-serif",
  xaxis_title_font_color = "#000000", yaxis_hidden = FALSE,
  yaxis_width = 120, yaxis_font_size = 11,
  yaxis_font_family = "sans-serif", yaxis_font_color = "#000000",
  yaxis_location = "right", yaxis_title = NULL,
  yaxis_title_font_size = 14, yaxis_title_font_family = "sans-serif",
  yaxis_title_font_color = "#000000", left_columns = NULL,
  left_columns_align = NULL, left_columns_font_size = 11,
  left_columns_font_family = "sans-serif",
  left_columns_font_color = "#000000", left_columns_subtitles = NULL,
  left_columns_subtitles_align = NULL, left_columns_subtitles_bold = NULL,
  left_columns_subtitles_font_size = 12,
  left_columns_subtitles_font_family = "sans-serif",
  left_columns_subtitles_font_color = "#000000", left_columns_title = NULL,
  left_columns_title_align = NULL, left_columns_title_bold = NULL,
  left_columns_title_font_size = 14,
  left_columns_title_font_family = "sans-serif",
  left_columns_title_font_color = "#000000", right_columns = NULL,
  right_columns_align = NULL, right_columns_font_size = 11,
  right_columns_font_family = "sans-serif",
  right_columns_font_color = "#000000", right_columns_title = NULL,
  right_columns_title_align = NULL, right_columns_title_bold = NULL,
  right_columns_title_font_size = 14,
  right_columns_title_font_family = "sans-serif",
  right_columns_title_font_color = "#000000",
  right_columns_subtitles = NULL, right_columns_subtitles_align = NULL,
  right_columns_subtitles_bold = NULL,
  right_columns_subtitles_font_size = 12,
  right_columns_subtitles_font_family = "sans-serif",
  right_columns_subtitles_font_color = "#000000", brush_color = "#0000FF",
  show_grid = TRUE, anim_duration = 500, table_style = FALSE, ...)
}
\arguments{
\item{x}{A numeric matrix
Defaults to \code{TRUE} unless \code{x} contains any \code{NA}s.}

\item{Rowv}{determines if and how the row dendrogram should be reordered.    By default, it is TRUE, which implies dendrogram is computed and reordered based on row means. If NULL or FALSE, then no dendrogram is computed and no reordering is done. If a dendrogram, then it is used "as-is", ie without any reordering. If a vector of integers, then dendrogram is computed and reordered based on the order of the vector.}

\item{Colv}{determines if and how the column dendrogram should be reordered.    Has the options as the Rowv argument above and additionally when x is a square matrix, Colv = "Rowv" means that columns should be treated identically to the rows.}

\item{distfun}{function used to compute the distance (dissimilarity) between both rows and columns. Defaults to dist.}

\item{hclustfun}{function used to compute the hierarchical clustering when Rowv or Colv are not dendrograms. Defaults to hclust.}

\item{dendrogram}{character string indicating whether to draw 'none', 'row', 'column' or 'both' dendrograms. Defaults to 'both'. However, if Rowv (or Colv) is FALSE or NULL and dendrogram is 'both', then a warning is issued and Rowv (or Colv) arguments are honoured.}

\item{reorderfun}{function(d, w) of dendrogram and weights for reordering the row and column dendrograms. The default uses stats{reorder.dendrogram}}

\item{k_row}{an integer scalar with the desired number of groups by which to color the dendrogram's branches in the rows (uses \link[dendextend]{color_branches})}

\item{k_col}{an integer scalar with the desired number of groups by which to color the dendrogram's branches in the columns (uses \link[dendextend]{color_branches})}

\item{symm}{logical indicating if x should be treated symmetrically; can only be true when x is a square matrix.}

\item{revC}{logical indicating if the column order should be reversed for plotting.
Default (when missing) - is FALSE, unless symm is TRUE.
This is useful for cor matrix.}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "none".}

\item{na.rm}{logical indicating whether NA's should be removed.}

\item{labRow}{character vectors with row labels to use (from top to bottom); default to rownames(x).}

\item{labCol}{character vectors with column labels to use (from left to right); default to colnames(x).}

\item{cexRow}{positive numbers. If not missing, it will override \code{xaxis_font_size}
and will give it a value cexRow*14}

\item{cexCol}{positive numbers. If not missing, it will override \code{yaxis_font_size}
and will give it a value cexCol*14}

\item{digits}{integer indicating the number of decimal places to be used by \link{round} for 'label'.}

\item{cellnote}{(optional) matrix of the same dimensions as \code{x} that has the human-readable version of each value, for displaying to the user on hover. If \code{NULL}, then \code{x} will be coerced using \code{\link{as.character}}.
If missing, it will use \code{x}, after rounding it based on the \code{digits} parameter.}

\item{cellnote_scale}{logical (default is FALSE). IF cellnote is missing and x is used,
should cellnote be scaled if x is also scaled?}

\item{cellnote_in_cell}{Overrides cellnote when the user wish to display texts inside the cells that are different from the tooltips.}

\item{show_cellnote_in_cell}{If \code{TRUE}, print cellnotes in the cells. Defaults to FALSE.}

\item{extra_tooltip_info}{A list of matrices that contains extra information to show in the tooltips. Dim of each matrix must equal to \code{x}.}

\item{color_range}{A vector of length 2 that specifies the range of values that colors get mapped to. The default is computed from \code{x}. If \code{x} is a factor, this argument is ignored.}

\item{lower_triangle}{A logical value to specify if only the lower triangle will be displayed. Defaults to FALSE and will give an error if \code{x} is not a square matrix.}

\item{cells_to_hide}{A boolean matrix with the same dimension as \code{x} that specifies which cell to hide. Hidden cells will not trigger tooltips and will be transparent in color.}

\item{cell_font_size}{Sets the maximum font size of cellnotes. Defauls to 11.}

\item{tip_font_size}{Sets the font size of texts in the tooltip. Defaults to 11.}

\item{show_legend}{logical. Defaults to TRUE. However, if \code{scale} is not "none", legend will not be shown.}

\item{legend_font_size}{positive integer. Sets the font size of the legend. Defaults to 11 (pixcels).}

\item{legend_width}{positive integer. Sets the desired width of the legend bar. Defaults to 60 (pixcels).}

\item{legend_digits}{positive integer. Sets the decimal places of the legend texts. Defaults to 1 if the max value of \code{x} is less than 1.0.}

\item{theme}{A custom CSS theme to use. Currently the only valid values are
\code{""} and \code{"dark"}. \code{"dark"} is primarily intended for
standalone visualizations, not R Markdown or Shiny.}

\item{colors}{Either a colorbrewer2.org palette name (e.g. \code{"YlOrRd"} or
\code{"Blues"}), or a vector of colors to interpolate in hexadecimal
\code{"#RRGGBB"} format, or a color interpolation function like
\code{\link[grDevices]{colorRamp}}.}

\item{width}{Width in pixels (optional, defaults to automatic sizing).}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}

\item{title}{character. Sets the title of the chart, defaults to NULL. The title is centred.}

\item{title_font_size}{integer. Font size of the chart title, defaults to 24 pixcels.}

\item{title_font_family}{character. Font family of the chart title, defaults to "sans-serif".}

\item{title_font_color}{An RGB character to set the color of the chart title. Defaults to "#000000".}

\item{subtitle}{character. Sets the subtitle of the chart, defaults to NULL. The subtitle is centred.}

\item{subtitle_font_size}{integer. Font size of the chart subtitle, defaults to 18 pixcels.}

\item{subtitle_font_family}{character. Font family of the chart subtitle, defaults to "sans-serif".}

\item{subtitle_font_color}{An RGB character to set the color of the chart subtitle. Defaults to "#000000".}

\item{footer}{character. Sets the footer of the chart, defaults to NULL. The footer is left-aligned.}

\item{footer_font_size}{integer. Font size of the chart footer_font_size, defaults to 11 pixcels.}

\item{footer_font_family}{character. Font family of the chart footer_font_family, defaults to "sans-serif".}

\item{footer_font_color}{An RGB character to set the color of the chart footer_font_color. Defaults to "#000000".}

\item{xaxis_hidden}{Boolean variable to hide x axis. Defaults to FALSE.}

\item{xaxis_height}{Size of axes, in pixels. Default is calculated automatically with axis font size specified.}

\item{xaxis_font_size}{Font size of axis labels, as a CSS size (e.g. "14px" or "12pt"). Defaults to 12.}

\item{xaxis_location}{Location of the x axis, c("bottom", "top"). Defaults to "bottom". "top" only works when dendrogram is "none".}

\item{xaxis_title}{Title text of the x axis.}

\item{xaxis_title_font_size}{Size of axis title text. Defaults to 14.}

\item{yaxis_hidden}{Boolean variable to hide y axis. Defaults to FALSE.}

\item{yaxis_width}{Size of axes, in pixels. Default is calculated automatically with axis font size specified.}

\item{yaxis_font_size}{Font size of axis labels, as a CSS size (e.g. "14px" or "12pt"). Defaults to 11.}

\item{yaxis_location}{Location of the y axis, c("right", "left"). Defaults to "right". "left" only works when dendrogram is "none".}

\item{yaxis_title}{Title text of the y axis.}

\item{yaxis_title_font_size}{Size of axis title text. Defaults to 14.}

\item{brush_color}{The base color to be used for the brush. The brush will be
filled with a low-opacity version of this color. \code{"#RRGGBB"} format
expected.}

\item{show_grid}{\code{TRUE} to show gridlines, \code{FALSE} to hide them, or
a numeric value to specify the gridline thickness in pixels (can be a
non-integer).}

\item{anim_duration}{Number of milliseconds to animate zooming in and out.
  For large \code{x} it may help performance to set this value to \code{0}.}

\item{...}{currently ignored}
}
\description{
Creates a D3.js-based heatmap widget. Modified from d3heatmap package.
}
\examples{
library(rhtmlHeatmap)
Heatmap(mtcars, scale = "column", colors = "Blues")
}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}

